+++ dk-club-subject/dk-club-infra/src/main/java/com/dk/subject/infra/basic/service/impl/SubjectLabelServiceImpl.java
@@ -23,7 +23,7 @@ public class SubjectLabelServiceImpl extends ServiceImpl<SubjectLabelMapper, Sub
     SubjectLabelMapper subjectLabelMapper;
 
     @Override
-    public SubjectLabel getSubjectLabel(Integer id) {
+    public SubjectLabel getSubjectLabel(Long id) {
         return subjectLabelMapper.selectById(id);
     }
